emergency/requirements.txt
flask
python-dotenv
chromadb
requests
twilio